test: reexec tagged product metrics profile
Replace the nested tagged gc build with the current tagged test binary re-exec and route the exact six-owner profile through required Linux CI plus existing local and macOS runners. Move testhook-only helper graphs behind the tag and lower the ordinary resource ratchets.\n\nThe focused process owner drops from 27.689s to 3.987s locally (about 6.9x faster).
