+++ test/test-resources.toml
@@ -10,7 +10,7 @@ version = 2
 [[audit_baseline]]
 scope = "all"
 resource = "subprocess"
-baseline_calls = 529
+baseline_calls = 528
 baseline_files = 161
 reported_calls = 495
 reported_files = 135
@@ -38,8 +38,8 @@ expires = "2026-10-01"
 [[debt]]
 scope = "untagged"
 resource = "subprocess"
-baseline_calls = 399
-baseline_files = 112
+baseline_calls = 393
+baseline_files = 111
 reported_calls = 380
 reported_files = 98
 owner_bead = "ga-80po0c.2"
@@ -64,8 +64,8 @@ expires = "2026-10-01"
 [[debt]]
 scope = "cmd/gc+untagged"
 resource = "environment"
-baseline_calls = 4338
-baseline_files = 205
+baseline_calls = 4336
+baseline_files = 204
 reported_calls = 3960
 reported_files = 184
 owner_bead = "ga-80po0c.2.3"
@@ -103,8 +103,8 @@ expires = "2026-10-01"
 [[debt]]
 scope = "untagged"
 resource = "http_test_server"
-baseline_calls = 318
-baseline_files = 67
+baseline_calls = 317
+baseline_files = 66
 reported_calls = 255
 reported_files = 56
 owner_bead = "ga-80po0c.2.2"
@@ -250,8 +250,8 @@ medium_reason = "package TestMain mutates process state"
 [[small_debt]]
 scope = "untagged"
 resource = "subprocess"
-baseline_calls = 396
-baseline_files = 110
+baseline_calls = 390
+baseline_files = 109
 reported_calls = 394
 reported_files = 105
 owner_bead = "ga-80po0c.2.1"
@@ -276,8 +276,8 @@ expires = "2026-10-01"
 [[small_debt]]
 scope = "cmd/gc+untagged"
 resource = "environment"
-baseline_calls = 4332
-baseline_files = 205
+baseline_calls = 4330
+baseline_files = 204
 reported_calls = 4348
 reported_files = 200
 owner_bead = "ga-80po0c.2.1"
@@ -315,8 +315,8 @@ expires = "2026-10-01"
 [[small_debt]]
 scope = "untagged"
 resource = "http_test_server"
-baseline_calls = 318
-baseline_files = 67
+baseline_calls = 317
+baseline_files = 66
 reported_calls = 300
 reported_files = 66
 owner_bead = "ga-80po0c.2.2"
